#27 - moves snack bar provider to ui module
